miscellaneous changes

+++ app/controllers/sessions_controller.rb
@@ -13,8 +13,15 @@ class SessionsController < ApplicationController
     end
 
     def destroy
-        session.delete(:user_id)
-        head: no_content
+        # session.delete(:user_id)
+        # head :no_content
+        user = User.find_by(id: session[:user_id])
+        if user
+            session.delete(:user_id)
+            head :no_content
+        else
+            render json: { errors: ["Not authorized"] }, status: :unauthorized
+        end
     end
 
 end
